# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DreamCar Reserva de vehiculos",
        "description": "Esta API nos permitira ver y manipular informacion a las reservas de los vehiculos.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rest/vehiculos/show/{id}": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Obtener un vehiculo por id",
                "description": "Renorna los detalles de un vehiculo",
                "operationId": "796a4253a51ad9ea5c413dd54eb18c31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehiculo",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del vehiculo con id seleccionado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "id_marca": {
                                            "type": "integer",
                                            "example": "99"
                                        },
                                        "modelo": {
                                            "type": "integer",
                                            "example": "2010"
                                        },
                                        "precio": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "disponible": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/vehiculos/vehiculo/{marca}": {
            "get": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Buscar vehiculo por una marca",
                "description": "Retorna vehiculos",
                "operationId": "983b2dde2f23eb77420949543ee90cfe",
                "parameters": [
                    {
                        "name": "marca",
                        "in": "path",
                        "description": "marca relacionada al vehiculo a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marca encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "vehiculo": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontro la marca del vehiculo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/vehiculos/vehiculo/{modelo}": {
            "post": {
                "tags": [
                    "Vehiculos"
                ],
                "summary": "Buscar vehiculo por modelo",
                "description": "Retorna un vehiculo",
                "operationId": "078255c81c980f73dc708963961d7493",
                "parameters": [
                    {
                        "name": "modelo",
                        "in": "path",
                        "description": "Modelo del vehiculo a buscar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehiculo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "marca": {
                                                "type": "string"
                                            },
                                            "modelo": {
                                                "type": "integer"
                                            },
                                            "precio": {
                                                "type": "integer"
                                            },
                                            "disponible": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No se encontro vehiculo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}